1. Spring Security Authentication 표준화 적용 		-UserDetailService, UserDetails 클래스 상속 클래스 구현 		-JwtAuthentication 클래스 표준화 적용 		-SecurityCOnfig의 접근 제한 적용(권한 체크 포함) 		-PostController, BoardController 		@AuthenticationPrincipal CustomerDetails 주입 테스트 		완료 		-PostController 보드별 게시글 리스트 가져오기 		-로그인 기능 Spring Security 표준화 적용

+++ src/main/java/com/example/board/post/PostService.java
@@ -67,10 +67,10 @@ public class PostService {
     /*    public PostDTO findById(Long id){
             return postRepository.findById(id).orElse(null);
         } */
-    @Transactional
-    public PostDTO getPost(Long postId) {
-        Post post = postRepository.findById(postId).orElseThrow(()->new NotFoundException((ErrorCode.POST_NOT_FOUND)));
-        return Post.toDTO(post);
+    @Transactional(readOnly = true)
+    public PostDTO getPost(Long loginUserId,Long id) {
+        Post post = postRepository.findById(id).orElseThrow(()->new NotFoundException((ErrorCode.POST_NOT_FOUND)));
+        return Post.toDTO(post, loginUserId);
     }
 
 
@@ -81,10 +81,12 @@ public class PostService {
         Post post = postRepository.findById(postId)
                         .orElseThrow(()->new NotFoundException(ErrorCode.POST_NOT_FOUND));
 
+        validateAuthor(post, loginUserId);
+
         //게시물 수정권한이 없다면 에러를 발생
         User user = post.getAuthor();
         if(!Objects.equals(user.getId(), loginUserId))
-            throw new ForbiddenException(ErrorCode.POST_ACCEPT_DENIED);
+            throw new ForbiddenException(ErrorCode.POST_ACCESS_DENIED);
 
         post.setTitle(request.getTitle());
         post.setBody(request.getBody());
@@ -100,11 +102,27 @@ public class PostService {
         Post post = postRepository.findById(postId)
                 .orElseThrow(()->new NotFoundException(ErrorCode.POST_NOT_FOUND));
 
+        validateAuthor(post, loginUserId);
+
         //게시물 수정권한이 없다면 에러를 발생
         User user = post.getAuthor();
         if(!Objects.equals(user.getId(), loginUserId))
-            throw new ForbiddenException(ErrorCode.POST_ACCEPT_DENIED);
+            throw new ForbiddenException(ErrorCode.POST_ACCESS_DENIED);
 
         postRepository.delete( post );
+    }
+
+    public void validateAuthor(Post post, Long userId){
+        if(!post.isAuthor(userId)){
+            throw new ForbiddenException(ErrorCode.POST_ACCESS_DENIED);
+        }
+    }
+
+    public List<PostDTO> findByBoardId(Long boardId) {
+       /* Board board = boardRepository.findById(boardId)
+                .orElseThrow(()->new NotFoundException(ErrorCode.BOARD_NOT_FOUND));*/
+
+        return postRepository.findByBoardId(boardId).stream()
+                .map(Post::toDTO).toList();
     }
 }
